requirements.txt
gradio>=4.0.0
Pillow>=9.0.0
pathlib 